text to weaver update + system prompt fixes
feature: added in a new endpoint which allows a prompt to definition functionality to occur. Good for experimentation and agent workflows.

fix: the prompting logic for the system prompt and array prompting so that information is correctly passed over and doesn't lead to blank values.

update: moved out a function which generates an object so that it can be used in multiple places inside of objectweaver. Additional benefit of having objectweaver be installable into other golang projects (think this will work)